+++ CHANGELOG.md
@@ -5,9 +5,9 @@
 ## v0.13.0
 #### Features and Improvements
 - Extension of the `data_ops` module:
+    - !161 extends the operations database with commonly used operations and makes operations on `np.`, `xr.` and `scipy.` modules easier.
     - !160 adds the `expression` operation for evaluating [`sympy`](https://www.sympy.org/en/index.html) expressions
     - !165 adds the `lambda` operation, which can be used to define callables, and the [`curve_fit`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.curve_fit.html) operation that requires such a callable.
-    - ...
 - !164 extends the [DAG minimal syntax](https://dantro.readthedocs.io/en/latest/data_io/transform.html#minimal-syntax) to allow specifying both positional and keyword arguments.
 - !160 adds [DAG syntax parsers](https://dantro.readthedocs.io/en/latest/data_io/dag_op_hooks.html) that can simplify DAG specification.
   Currently, this includes only a parser for the `expression` operation (#149).
@@ -20,6 +20,7 @@
 
 ## v0.12.5
 - !159 fixes a bug (#147) which led to duplicate DAG cache files after a storage function error.
+
 
 ## v0.12.4
 - !151 adds documentation of the `GraphGroup`.